+++ tests/apps/test_app/management/commands/field_value.py
@@ -3,11 +3,18 @@ from typing import Annotated
 import typer
 from django.utils.translation import gettext_lazy as _
 from datetime import timedelta
+from django.core.management.base import CommandError
 
 from django_typer.management import TyperCommand, model_parser_completer
 from django_typer.parsers.model import ReturnType
 from tests.apps.test_app.models import ShellCompleteTester
 from django_typer.utils import duration_iso_string
+
+
+def test_custom_error_message(model_cls, field_value: str, exception: Exception):
+    raise CommandError(
+        f"Test custom error message: {model_cls=}, {field_value=}, {exception=}"
+    )
 
 
 class Command(TyperCommand, rich_markup_mode="rich"):
@@ -17,7 +24,10 @@ class Command(TyperCommand, rich_markup_mode="rich"):
             timedelta,
             typer.Argument(
                 **model_parser_completer(
-                    ShellCompleteTester, return_type=ReturnType.FIELD_VALUE
+                    ShellCompleteTester,
+                    lookup_field="duration_field",
+                    return_type=ReturnType.FIELD_VALUE,
+                    on_error=test_custom_error_message,
                 )
             ),
         ],