+++ doc/source/changelog.rst
@@ -9,6 +9,7 @@ v3.0.0 (202X-XX-XX)
 
 * Implemented `Migrate pyproject.toml to poetry 2 and portable project specifiers. <https://github.com/django-commons/django-typer/issues/164>_`
 * BREAKING `Split parsers.py and completers.py into submodules. <https://github.com/django-commons/django-typer/issues/163>_`
+* Implemented `Model completer/parser should support returning the field value <https://github.com/django-commons/django-typer/issues/162>`_
 * Fixed `Model objects with null lookup fields should not be included in model field completion output <https://github.com/django-commons/django-typer/issues/160>`_
 * Implemented `Add security scans to CI. <https://github.com/django-commons/django-typer/issues/158>`_
 * Implemented `Add a performance regression. <https://github.com/django-commons/django-typer/issues/157>`_
@@ -26,6 +27,7 @@ v3.0.0 (202X-XX-XX)
 * Fixed `Remove management imports in django_typer/__init__.py <https://github.com/django-commons/django-typer/issues/95>`_
 * Fixed `ParamSpec includes self for group methods <https://github.com/django-commons/django-typer/issues/73>`_
 * Fixed `Installed shellcompletion scripts do not pass values of --settings or --pythonpath <https://github.com/django-commons/django-typer/issues/68>`_
+* Implemented `Add support for QuerySet parameter types. <https://github.com/django-commons/django-typer/issues/58>`_
 * Fixed `shellcompletion complete should print to the command's stdout. <https://github.com/django-commons/django-typer/issues/19>`_
 * Implemented `Add completer/parser for FileField and FilePathField <https://github.com/django-commons/django-typer/issues/17>`_
 * Implemented `Add completer/parser for DurationField <https://github.com/django-commons/django-typer/issues/16>`_